[Race condtion] Try to fix race condition part 2
