+++ actor.go
@@ -1,9 +1,82 @@
 package main
 
 import (
+	"github.com/tin-incognito/simulate-consensus/common"
 	"github.com/tin-incognito/simulate-consensus/utils"
+	"log"
+	"sync"
 	"time"
 )
+
+//Actor
+type Actor struct{
+	PrePrepareMsgCh chan NormalMsg
+	PrepareMsgCh    chan NormalMsg
+	CommitMsgCh chan NormalMsg
+	ViewChangeMsgCh chan ViewMsg
+	BroadcastMsgCh chan bool
+	isStarted      bool
+	StopCh         chan struct{}
+	CurrNode 	   *Node
+	ProposalNode *Node
+	Validators map[int]*Node
+	chainHandler ChainHandler
+	BFTMsgLogs map[string]*NormalMsg
+	ViewChangeMsgLogs map[string]*ViewMsg
+	timeOutCh chan bool
+	errCh chan error
+	wg sync.WaitGroup
+	postAmountMsgTimerCh chan bool
+	postMsgTimerCh chan MsgTimer
+	prepareAmountMsgTimer *time.Timer
+	commitAmountMsgTimer *time.Timer
+	idleTimer *time.Timer
+	blockPublishTimer *time.Timer
+	commitTimer *time.Timer
+	viewChangeTimer *time.Timer
+	newBlock *Block
+	prePrepareMsg map[int]*NormalMsg
+	viewChangeExpire bool
+	viewChangeAmount map[int]int
+	phaseStatus string
+	msgTimerCh chan MsgTimer
+	stuckCh chan string
+}
+
+func NewActor() *Actor{
+
+	res := &Actor{
+		PrePrepareMsgCh:     make (chan NormalMsg, 100),
+		PrepareMsgCh:        make (chan NormalMsg, 100),
+		CommitMsgCh:         make (chan NormalMsg, 100),
+		ViewChangeMsgCh: make (chan ViewMsg, 100),
+		BroadcastMsgCh:      make (chan bool, 100),
+		isStarted:           true,
+		CurrNode:            nil,
+		ProposalNode: nil,
+		Validators: make(map[int]*Node),
+		StopCh: make(chan struct{}, 20),
+		chainHandler: &Chain{},
+		BFTMsgLogs: make(map[string]*NormalMsg),
+		ViewChangeMsgLogs: make(map[string]*ViewMsg),
+		wg:                  sync.WaitGroup{},
+		timeOutCh: make(chan bool, 100),
+		errCh: make(chan error, 100),
+		viewChangeAmount: make(map[int]int),
+		msgTimerCh: make (chan MsgTimer),
+		postAmountMsgTimerCh: make (chan bool),
+		prePrepareMsg: make (map[int]*NormalMsg),
+		stuckCh: make (chan string),
+	}
+
+	return res
+}
+
+//Name return name of actor to user
+func (actor Actor) Name() (string,error){
+	return common.SawToothConsensus, nil
+}
+
 
 func (actor *Actor) updateAllViewChangeMode(){
 	for _, element := range actor.Validators{
@@ -167,4 +240,74 @@ func (engine *Engine) start() error{
 
 func (actor *Actor) View() uint64{
 	return actor.CurrNode.View
+}
+
+//handleMsgTimer ...
+func (actor *Actor) handleMsgTimer(msgTimer MsgTimer){
+
+	log.Println("actor.prepareAmountMsgTimer:", actor.prepareAmountMsgTimer)
+
+	switch msgTimer.Type {
+	case PREPARE:
+
+		go func(){
+			select {
+			case <- actor.prepareAmountMsgTimer.C:
+				currActor := actor.CurrNode.consensusEngine.BFTProcess
+
+				log.Println(1)
+
+				prepareMutex.Lock()
+
+				prePrepareMsg := currActor.prePrepareMsg[int(currActor.CurrNode.index)]
+
+				for _, msg := range currActor.BFTMsgLogs{
+					if msg.prevMsgHash != nil && *msg.prevMsgHash == prePrepareMsg.hash && msg.Type == PREPARE {
+						currActor.BFTMsgLogs[prePrepareMsg.hash].Amount++
+					}
+				}
+
+				if uint64(currActor.BFTMsgLogs[prePrepareMsg.hash].Amount) <= uint64(2*n/3) {
+
+					//switchViewChangeModeMutex.Lock()
+					//currActor.switchToviewChangeMode()
+					//switchViewChangeModeMutex.Unlock()
+
+					return
+				}
+
+				if !currActor.BFTMsgLogs[prePrepareMsg.hash].prepareExpire{
+					//Move to committing phase
+					msg := NormalMsg{
+						hash: 	   utils.GenerateHashV1(),
+						Type:      COMMIT,
+						View:      currActor.chainHandler.View(),
+						SeqNum:    currActor.chainHandler.SeqNumber(),
+						SignerID:  currActor.CurrNode.index,
+						Timestamp: uint64(time.Now().Unix()),
+						BlockID:   prePrepareMsg.BlockID,
+						block: prePrepareMsg.block,
+						prevMsgHash: &prePrepareMsg.hash,
+					}
+
+					for _, member := range currActor.Validators{
+						go func(node *Node){
+							log.Println("Send to commit channel")
+							node.consensusEngine.BFTProcess.CommitMsgCh <- msg
+						}(member)
+					}
+					currActor.BFTMsgLogs[prePrepareMsg.hash].prepareExpire = true
+				}
+
+				currActor.postAmountMsgTimerCh <- true
+				prepareMutex.Unlock()
+
+			}
+		}()
+
+	case COMMIT:
+
+	case VIEWCHANGE:
+
+	}
 }
