+++ timer.go
@@ -2,5 +2,6 @@ package main
 
 //MsgTimer ...
 type MsgTimer struct{
+	//msg NormalMsg
 	Type string
 }
